Add paint packs to 3dModeler with model file v4 format
Paint palette is no longer fixed at 16 colors; each model carries N packs of
16, switchable via a Paint Pack dropdown above the color wheel (with add and
delete). Triangle paint indices widen to int16_t and store the global
(pack*16+slot) so painted triangles keep their color when the user flips packs.
.mdl bumps to v4 (variable palette + 2-byte triColors); v3 files still load.

+++ beta-2.0/VisualEngine/inputManagement/Collision.h
@@ -20,7 +20,7 @@ struct BlockCollider {
     std::vector<Triangle> triangles;  // empty for rectangular meshes (uses AABB only)
     std::string meshName;
     bool isRectangular;               // true = AABB-only, false = triangle-level
-    std::vector<int8_t> triColors;    // palette index per triangle (-1 = unpainted)
+    std::vector<int16_t> triColors;   // global palette index per triangle (pack*16+slot, -1 = unpainted)
 };
 
 struct CollisionHit {
